Solution

+++ app/main.py
@@ -1,12 +1,15 @@
 import os
 
 
-def move_file(command: str) -> None:
+def move_file(command: str) -> None | str:
     files = command.split(" ")
-    for i in range(len(files[2].split("/")[:-1])):
-        if not os.path.exists("/".join(files[2].split("/")[:i + 1])):
-            os.mkdir("/".join(files[2].split("/")[:i + 1]))
-    with open(files[2], "w") as f2:
-        with open(files[1], "r") as f1:
-            f2.write(f1.read())
-        os.remove(files[1])
+    if len(files) != 3 or files[0] != "mv":
+        return "Invalid command"
+    source_file = files[1]
+    destination_file = files[2]
+    if len(destination_file.split("/")) > 1:
+        os.makedirs("/".join(destination_file.split("/")[:-1]), exist_ok=True)
+    with open(destination_file, "w") as new_file, \
+         open(source_file, "r") as old_file:
+        new_file.write(old_file.read())
+    os.remove(files[1])
